requirements.txt
python-dotenv
streamlit
langchain_community
langchain
langchain_google_genai
docx
langchain-huggingface